GETP-216 refactor: Email의 JPA에 대한 의존성을 완전히 제거
