+++ src/main/java/es/princip/getp/api/controller/auth/SignUpController.java
@@ -5,7 +5,7 @@ import es.princip.getp.api.controller.auth.dto.request.SignUpRequest;
 import es.princip.getp.api.controller.common.dto.ApiResponse;
 import es.princip.getp.api.controller.common.dto.ApiResponse.ApiSuccessResult;
 import es.princip.getp.application.auth.service.SignUpService;
-import es.princip.getp.domain.member.model.Email;
+import es.princip.getp.domain.common.model.Email;
 import jakarta.validation.Valid;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.HttpStatus;
