+++ src/main/java/es/princip/getp/api/controller/people/command/dto/request/RegisterPeopleRequest.java
@@ -4,8 +4,8 @@ import es.princip.getp.api.validation.EmailPattern;
 import es.princip.getp.api.validation.Enum;
 import es.princip.getp.api.validation.PhoneNumberPattern;
 import es.princip.getp.application.people.command.RegisterPeopleCommand;
+import es.princip.getp.domain.common.model.Email;
 import es.princip.getp.domain.common.model.PhoneNumber;
-import es.princip.getp.domain.member.model.Email;
 import es.princip.getp.domain.member.model.Nickname;
 import es.princip.getp.domain.people.model.PeopleType;
 import jakarta.validation.constraints.NotBlank;
