+++ src/main/java/es/princip/getp/application/auth/service/VerificationService.java
@@ -3,7 +3,7 @@ package es.princip.getp.application.auth.service;
 import es.princip.getp.application.auth.exception.IncorrectVerificationCodeException;
 import es.princip.getp.application.auth.exception.NotFoundVerificationException;
 import es.princip.getp.common.util.RandomUtil;
-import es.princip.getp.domain.member.model.Email;
+import es.princip.getp.domain.common.model.Email;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
