+++ src/main/java/es/princip/getp/application/client/command/EditClientCommand.java
@@ -2,9 +2,9 @@ package es.princip.getp.application.client.command;
 
 import es.princip.getp.domain.client.model.Address;
 import es.princip.getp.domain.client.model.BankAccount;
-import es.princip.getp.domain.member.model.Email;
-import es.princip.getp.domain.member.model.Nickname;
+import es.princip.getp.domain.common.model.Email;
 import es.princip.getp.domain.common.model.PhoneNumber;
+import es.princip.getp.domain.member.model.Nickname;
 
 public record EditClientCommand(
     Long memberId,
