+++ src/main/java/es/princip/getp/application/client/command/RegisterClientCommand.java
@@ -2,10 +2,10 @@ package es.princip.getp.application.client.command;
 
 import es.princip.getp.domain.client.model.Address;
 import es.princip.getp.domain.client.model.BankAccount;
-import es.princip.getp.domain.member.model.Email;
+import es.princip.getp.domain.common.model.Email;
+import es.princip.getp.domain.common.model.PhoneNumber;
 import es.princip.getp.domain.member.model.Member;
 import es.princip.getp.domain.member.model.Nickname;
-import es.princip.getp.domain.common.model.PhoneNumber;
 
 public record RegisterClientCommand(
     Member member,
