+++ src/main/java/es/princip/getp/application/people/command/RegisterPeopleCommand.java
@@ -1,7 +1,7 @@
 package es.princip.getp.application.people.command;
 
+import es.princip.getp.domain.common.model.Email;
 import es.princip.getp.domain.common.model.PhoneNumber;
-import es.princip.getp.domain.member.model.Email;
 import es.princip.getp.domain.member.model.Nickname;
 import es.princip.getp.domain.people.model.PeopleType;
