+++ src/main/java/es/princip/getp/domain/people/model/PeopleInfo.java
@@ -1,7 +1,7 @@
 package es.princip.getp.domain.people.model;
 
 import es.princip.getp.domain.common.model.BaseModel;
-import es.princip.getp.domain.member.model.Email;
+import es.princip.getp.domain.common.model.Email;
 import jakarta.validation.constraints.NotNull;
 import lombok.Getter;
