+++ src/main/java/es/princip/getp/persistence/adapter/member/MemberPersistenceAdapter.java
@@ -4,7 +4,7 @@ import es.princip.getp.application.member.port.out.CheckMemberPort;
 import es.princip.getp.application.member.port.out.LoadMemberPort;
 import es.princip.getp.application.member.port.out.SaveMemberPort;
 import es.princip.getp.application.member.port.out.UpdateMemberPort;
-import es.princip.getp.domain.member.model.Email;
+import es.princip.getp.domain.common.model.Email;
 import es.princip.getp.domain.member.model.Member;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Repository;
