+++ src/test/java/es/princip/getp/api/controller/auth/SignUpControllerTest.java
@@ -18,7 +18,7 @@ import java.util.Set;
 
 import static es.princip.getp.api.docs.FieldDescriptorHelper.getDescriptor;
 import static es.princip.getp.fixture.auth.EmailVerificationFixture.VERIFICATION_CODE;
-import static es.princip.getp.fixture.member.EmailFixture.EMAIL;
+import static es.princip.getp.fixture.common.EmailFixture.EMAIL;
 import static es.princip.getp.fixture.member.PasswordFixture.PASSWORD;
 import static org.mockito.ArgumentMatchers.any;
 import static org.mockito.BDDMockito.given;
