+++ src/test/java/es/princip/getp/api/controller/client/query/MyClientQueryControllerTest.java
@@ -18,7 +18,7 @@ import java.time.LocalDateTime;
 import static es.princip.getp.api.docs.HeaderDescriptorHelper.authorizationHeaderDescriptor;
 import static es.princip.getp.fixture.client.AddressFixture.address;
 import static es.princip.getp.fixture.client.BankAccountFixture.bankAccount;
-import static es.princip.getp.fixture.member.EmailFixture.EMAIL;
+import static es.princip.getp.fixture.common.EmailFixture.EMAIL;
 import static es.princip.getp.fixture.member.NicknameFixture.NICKNAME;
 import static es.princip.getp.fixture.member.PhoneNumberFixture.PHONE_NUMBER;
 import static es.princip.getp.fixture.member.ProfileImageFixture.profileImage;
