+++ src/test/java/es/princip/getp/api/controller/people/command/MyPeopleControllerTest.java
@@ -22,7 +22,7 @@ import static es.princip.getp.api.docs.HeaderDescriptorHelper.authorizationHeade
 import static es.princip.getp.api.docs.PayloadDocumentationHelper.responseFields;
 import static es.princip.getp.domain.member.model.MemberType.ROLE_CLIENT;
 import static es.princip.getp.domain.member.model.MemberType.ROLE_PEOPLE;
-import static es.princip.getp.fixture.member.EmailFixture.EMAIL;
+import static es.princip.getp.fixture.common.EmailFixture.EMAIL;
 import static es.princip.getp.fixture.member.NicknameFixture.NICKNAME;
 import static es.princip.getp.fixture.member.PhoneNumberFixture.PHONE_NUMBER;
 import static org.mockito.ArgumentMatchers.any;
