+++ src/test/java/es/princip/getp/api/security/annotation/WithCustomMockUser.java
@@ -7,7 +7,7 @@ import java.lang.annotation.Retention;
 import java.lang.annotation.RetentionPolicy;
 
 import static es.princip.getp.domain.member.model.MemberType.ROLE_PEOPLE;
-import static es.princip.getp.fixture.member.EmailFixture.EMAIL;
+import static es.princip.getp.fixture.common.EmailFixture.EMAIL;
 import static es.princip.getp.fixture.member.NicknameFixture.NICKNAME;
 import static es.princip.getp.fixture.member.PasswordFixture.PASSWORD;
 import static es.princip.getp.fixture.member.PhoneNumberFixture.PHONE_NUMBER;
