+++ src/test/java/es/princip/getp/api/security/annotation/WithCustomMockUserSecurityContextFactory.java
@@ -1,6 +1,7 @@
 package es.princip.getp.api.security.annotation;
 
 import es.princip.getp.api.security.details.PrincipalDetails;
+import es.princip.getp.domain.common.model.Email;
 import es.princip.getp.domain.common.model.PhoneNumber;
 import es.princip.getp.domain.member.model.*;
 import org.springframework.security.authentication.UsernamePasswordAuthenticationToken;
