+++ src/test/java/es/princip/getp/fixture/member/MemberFixture.java
@@ -1,13 +1,13 @@
 package es.princip.getp.fixture.member;
 
-import es.princip.getp.domain.member.model.Email;
+import es.princip.getp.domain.common.model.Email;
 import es.princip.getp.domain.member.model.Member;
 import es.princip.getp.domain.member.model.MemberType;
 
 import java.util.List;
 import java.util.stream.IntStream;
 
-import static es.princip.getp.fixture.member.EmailFixture.email;
+import static es.princip.getp.fixture.common.EmailFixture.email;
 
 public class MemberFixture {
